Replace piece theme.
I'm not sure if I could be using the merida pieces, I noticed the
license according to lichess copying.md is the GPL (which I wasn't
properly attributing). But to be on the safe side I switched to the
cburnett pieces which can be used under CC-BY-SA and have a more
clear authorship/license page.