Fix the mateIn3 search and improve the themes list
Using a separate function to get the db theme tags, so I can format
the names in the list nicely.

Make the search slightly less dumb, there was an unneeded
"to_lowercase" there.

Fix the mateIn3 tag, I had made a mistake while copying the themes
and it was searching for mateIn2 instead.