Update to iced 0.6 and tweak/add some color themes
Now the glow backend support images, so we don't need to use wgpu
anymore (finally no more need for 2 executables!).
The styling in iced is totally different now, but it made the code
more clear and it's quite easy now to add more color schemes if
needed.
For now we had to switch to iced_aw git instead of the crates io
version, since they didn't release a version with iced 0.6 support
yet.