Add search by opening variation
I noticed that the tags in the lichess puzzle dashboard don't seem
to make complete sense, there's duplicity and the "opening family"
doesn't look that well organized, which can make the search a bit
confusing, but anyway, it's better to have the same filters as
lichess.

I didn't add all the variations though, only the ones with more
than 1k puzzles tagged and a few other exceptions.