Add 3 piece sets (mainly Merida & Alpha).
I was avoiding non-free piece sets, but since we're already using
the Alpha font (to export to PDF there's basically no libre
option), there's no point in not including the SVG pieces too.

And since we're at it, I added Merida too, since it's probably the
most popular set of them all. And after all, despite being non-free,
I think it's in the spirit of the authors' intentions to use them in
a free app like this one.

I also added Governor, another of sadsnake's sets, this one is
creative commons and has a very elegant design.